Stop provisioning on errors
This causes the ansible playbook to stop immediately if any host has an
error on a task instead of the default where it continues with the
remaining (non-error) hosts. The reason for this change is that a
cluster that is missing some hosts isn't very useful, so we might as
well stop.

Signed-off-by: John Strunk <jstrunk@redhat.com>